feat(coin-pools): BigVector order storage + encoded u128 order ids
Coin pools (`Pool<Base, Quote>`) now store orders in a `BigVector<Order>`
keyed by a DeepBook-style encoded `u128` order id, instead of a sorted
`vector<Order>` with opaque `u64` serials. Multicoin pools keep the vector
book and serial ids permanently; the two stacks run side by side.

Why: the vector book rewrote the whole bids-or-asks vector into the pool
object on every placement, which grows the object without bound and caps
book depth at Sui's max object size. Keying by order id also makes cancel,
modify and lookup O(log n) instead of a linear scan plus memmove.

Wherever the two trading types conflicted, the shared module is forked
rather than widened, with the coin copy `coin_`-prefixed:

  coin_book/{coin_book,coin_order,coin_order_info,coin_fill}.move
  state/{coin_account,coin_state}.move
  vault/coin_vault.move
  coin_order_query.move            (replaces order_query.move)
  helper/big_vector.move           (vendored verbatim from deepbookv3)

`vault.move` keeps only the fee-event helpers `multicoin_vault` shares.
`helper/utils.move` gains the order-id codec; `constants.move` gains the
two per-side sequence seeds. The order id encodes side in bit 127 and
price in bits 64..126, and the per-side counters run in opposite
directions, so key order alone yields price-time priority on both sides.

Multicoin is untouched: `multicoin_pool.move`, `multicoin_vault.move`,
`book/{order,order_info,fill}.move`, `state/{account,state}.move` and
every multicoin test are byte-identical to cycle-7, and all 183 multicoin
tests pass unmodified. `book.move`'s only change is comments — 294 lines
of commented-out BigVector code deleted, now that it is real code in
coin_book.

Behaviour changes on the coin side, both asserted in tests:
- an absent order id aborts with `big_vector::ENotFound` rather than a
  book-level code, because the id is the storage key and the book never
  searches for it;
- `iter_orders` with a `start_order_id` naming no live order returns an
  empty page instead of silently re-serving page one.

Gas: BigVector costs ~8-17% more computation at every depth the benchmark
suite reaches (depth 300: +7.7%), with no crossover in range — 64-order
slices mean depth 300 is only ~5 slices, and the unit-test meter models
computation but not the storage cost this change exists to remove. See
docs/plans/coin-pool-bigvector-side-by-side.md §9 for the full table.

617 tests pass (8 new slice-boundary depth tests, 21 vendored BigVector
tests); package warning count 107 -> 104.

+++ packages/triex/tests/pool_tests.move
@@ -9,7 +9,7 @@
 #[test_only]
 module triex::pool_tests {
     use sui::{object::ID, test_scenario::Scenario};
-    use triex::{order_info::OrderInfo, pool_test_utils};
+    use triex::{coin_order_info::OrderInfo, pool_test_utils};
 
     public fun setup_everything<BaseAsset, QuoteAsset, ReferenceBaseAsset, ReferenceQuoteAsset>(
         test: &mut Scenario,
@@ -152,7 +152,7 @@ module triex::pool_tests {
         sender: address,
         pool_id: ID,
         trading_account_id: ID,
-        order_id: u64,
+        order_id: u128,
         test: &mut Scenario,
     ) {
         pool_test_utils::cancel_order<BaseAsset, QuoteAsset>(
@@ -174,7 +174,7 @@ module triex::pool_tests {
         sender: address,
         pool_id: ID,
         trading_account_id: ID,
-        order_id: u64,
+        order_id: u128,
         new_quantity: u64,
         test: &mut Scenario,
     ) {
